# api/openapi.yaml
openapi: 3.0.3
info:
  title: QLNS API
  version: 1.0.0
  description: >-
    Contract-first API for Recruitment, Core HR and Contract Management.
    Workflow changes use explicit commands, ETag/If-Match concurrency control,
    server-side authorization, audit logging and transactional outbox delivery.
servers:
  - url: http://localhost:5000
    description: Local ASP.NET Core development server
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/recruitment/applications/{applicationId}:
    get:
      tags: [Recruitment Pipeline]
      operationId: getRecruitmentApplication
      summary: Get one authorized recruitment application
      x-requirement: REC-03.1, REC-03.2
      parameters: [{ $ref: '#/components/parameters/ApplicationId' }]
      responses:
        '200':
          description: Application
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/RecruitmentApplication' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/recruitment/applications/{applicationId}/advance:
    post:
      tags: [Recruitment Pipeline]
      operationId: advanceRecruitmentApplication
      summary: Advance an application exactly one active stage
      x-requirement: REC-03.2
      parameters: [{ $ref: '#/components/parameters/ApplicationId' }, { $ref: '#/components/parameters/IfMatch' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdvanceApplicationRequest' } } } }
      responses:
        '200':
          description: Application advanced
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content: { application/json: { schema: { $ref: '#/components/schemas/RecruitmentApplication' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: OIDC access token for workforce users. }
    offerToken: { type: apiKey, in: header, name: X-Offer-Token, description: Short-lived single-offer candidate token. }
  parameters:
...
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^"[1-9][0-9]*"$' }, example: '"4"' }
...
    ApplicationId: { name: applicationId, in: path, required: true, schema: { type: integer, format: int64, minimum: 1 } }
...
  headers:
    ETag: { description: Quoted numeric resource version., schema: { type: string, example: '"5"' } }
...
  responses:
    BadRequest: { description: Malformed request or If-Match header, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    Unauthorized: { description: Authentication is required or invalid, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    Forbidden: { description: Insufficient permission, field access or data scope, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    NotFound: { description: Resource is not visible or does not exist, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ProblemDetails' } } } }
    Conflict: { description: Workflow, prerequisite, uniqueness or concurrency conflict, content: { application/problem+json: { schema: { $ref: '#/components/schemas/BusinessProblemDetails' } } } }
    UnprocessableEntity: { description: Semantically invalid fields, content: { application/problem+json: { schema: { $ref: '#/components/schemas/ValidationProblemDetails' } } } }
...
  schemas:
...
    ProblemDetails:
      type: object
      additionalProperties: true
      required: [type, title, status, correlationId]
      properties: { type: { type: string, format: uri }, title: { type: string }, status: { type: integer, minimum: 400, maximum: 599 }, detail: { type: string }, instance: { type: string }, correlationId: { type: string }, code: { type: string } }
    BusinessProblemDetails:
      allOf: [{ $ref: '#/components/schemas/ProblemDetails' }, { type: object, required: [code], properties: { code: { type: string } } }]
    ValidationProblemDetails:
      allOf: [{ $ref: '#/components/schemas/ProblemDetails' }, { type: object, required: [errors], properties: { errors: { type: object, additionalProperties: { type: array, items: { type: string } } } } }]
...
    RecruitmentStage: { type: string, enum: [sourced_applied, ai_screening, tech_interview, executive_round, offer_letter, hired_ready, rejected, withdrawn] }
    ActiveRecruitmentStage: { type: string, enum: [sourced_applied, ai_screening, tech_interview, executive_round, offer_letter, hired_ready] }
    RecruitmentApplication:
      type: object
      required: [id, candidateId, jobPostingId, stage, source, appliedAt, version, updatedAt]
      properties: { id: { type: integer, format: int64 }, candidateId: { type: integer, format: int64 }, jobPostingId: { type: integer, format: int64 }, resumeId: { type: integer, format: int64, nullable: true }, stage: { $ref: '#/components/schemas/RecruitmentStage' }, aiScore: { type: number, minimum: 0, maximum: 100, nullable: true }, source: { type: string }, appliedAt: { type: string, format: date-time }, version: { type: integer, format: int64, minimum: 1 }, updatedAt: { type: string, format: date-time } }
...
    AdvanceApplicationRequest: { type: object, additionalProperties: false, required: [targetStage], properties: { targetStage: { $ref: '#/components/schemas/ActiveRecruitmentStage' }, reason: { type: string, maxLength: 1000, nullable: true } } }
...